blk_archive: Add common.py file, move BlkArchive class to it
Code sharing between different blk-archive tests.
