+++ src/dmtest/blk_archive/rolling_snaps.py
@@ -16,37 +16,9 @@ import os
 import threading
 import logging as log
 
+from dmtest.blk_archive.common import BlkArchive
+
 # --------------------------------
-
-
-class BlkArchive:
-    def __init__(self, dir, block_size=4096):
-        self.dir = os.path.abspath(dir)
-        process.run(f"blk-archive create -a {self.dir} --block-size {block_size}")
-
-    def pack(self, dev):
-        process.run(f"blk-archive pack -a {self.dir} {dev}")
-        return self.get_stream_id(dev)
-
-    def pack_delta(self, old, old_id, new):
-        process.run(f"blk-archive pack -a {self.dir} {new} --delta-stream {old_id} --delta-device {old}")
-        return self.get_stream_id(new)
-
-    def verify(self, dev):
-        process.run(
-            f"blk-archive verify -a {self.dir} --stream {self.get_stream_id(dev)} {dev}"
-        )
-
-    def get_stream_id(self, dev):
-        name = os.path.basename(dev.path)
-        (code, stdout, stderr) = process.run(f"blk-archive list -a {self.dir}")
-        grep_output = [line for line in stdout.split("\n") if name in line]
-        if len(grep_output) != 1:
-            raise ValueError(f"couldn't find stream for {name}")
-
-        # Run cut equivalent
-        cut_output = [line.split(" ")[0] for line in grep_output]
-        return cut_output[0]
 
 
 def t_rolling_snaps(fix):
